in coach.py in the method get_response we might respond with a message, depending on the current distance_round_track. Currently we loop through all the messages and pick the one that is closest to the distance. In order to only respond once to a distance, we memorize the time we responded in the 'read' entry. Refactor this code, so that it does not depend on the time read of the message.
Maybe order the messages in a stack, ordered by distance, which is in the 'at' key of a message.
Then only look at top message and compare it to the distance.
Once close enough respond, if all criteria is met.
Then move the message to the bottom of the stack.
This means self.msg_read_interval must not be used
and msg['read'] must not be initialized or used.

The initialization of the stack should be in its own method with a distance as the argument.
It would return a stack with message at the top which is closest to the argument.

We still want to get the parameter `now` in `notify` and `get_response`
because we want to pass it to history.update - so dont remove that.

dont remove code that is commented out and dont remove methods that are not used.